+++ fastapi_cache/decorator.py
@@ -96,8 +96,9 @@ def cache(
             copy_kwargs = kwargs.copy()
             request = copy_kwargs.pop("request", None)
             response = copy_kwargs.pop("response", None)
+
             if (
-                request and request.headers.get("Cache-Control") == "no-store"
+                request and request.headers.get("Cache-Control") in ("no-store", "no-cache")
             ) or not FastAPICache.get_enable():
                 return await ensure_async_func(*args, **kwargs)