+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 
 ## 0.1
 
+### 0.1.10
+
+- Add `Cache-Control:no-cache` support.
+
 ### 0.1.9
 
 - Replace `aioredis` with `redis-py`.